# docs/swagger.yaml
definitions:
  types.Book:
    properties:
      description:
        type: string
      id:
        type: integer
      isbn:
        type: string
      name:
        type: string
      title:
        type: string
    type: object
  types.BookCreateInput:
    properties:
      description:
        type: string
      isbn:
        type: string
      name:
        type: string
      title:
        type: string
    required:
    - description
    - isbn
    - name
    - title
    type: object
  types.BookUpdateInput:
    properties:
      description:
        type: string
      isbn:
        type: string
      name:
        type: string
      title:
        type: string
    type: object
  types.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  types.SignInInput:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  types.SignUpInput:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    required:
    - email
    - name
    - password
    type: object
info:
  contact: {}
  description: API Server for Library Application
  title: Library App API
  version: "1.0"
paths:
  /api/v1/books/:
    get:
      operationId: get-all-books
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.Book'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Get all
      tags:
      - book
    post:
      operationId: create-book
      parameters:
      - description: ' '
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/types.BookCreateInput'
      responses:
        "201":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Create
      tags:
      - book
  /api/v1/books/{id}:
    delete:
      operationId: delete-book-by-id
      parameters:
      - description: ' '
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Delete by id
      tags:
      - book
    get:
      description: get book by id
      operationId: get-book-by-id
      parameters:
      - description: ' '
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Book'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Get by id
      tags:
      - book
    put:
      operationId: update-book-by-id
      parameters:
      - description: ' '
        in: path
        name: id
        required: true
        type: string
      - description: ' '
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/types.BookUpdateInput'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Update by id
      tags:
      - book
  /api/v1/users/auth/log-out:
    get:
      operationId: log-out
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Log out
      tags:
      - auth
  /api/v1/users/auth/sign-in:
    post:
      operationId: sign-in
      parameters:
      - description: ' '
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/types.SignInInput'
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Sign in
      tags:
      - auth
  /api/v1/users/auth/sign-up:
    post:
      operationId: sign-up
      parameters:
      - description: ' '
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/types.SignUpInput'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Sign up
      tags:
      - auth
swagger: "2.0"